+++ contracts/VoterV3.sol
@@ -291,7 +291,7 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
         emit RemovedFromWhitelist(msg.sender, _token);
     }
 
-     /// @notice Kill a malicious gauge 
+    /// @notice Kill a malicious gauge 
     /// @param  _gauge gauge to kill
     function killGauge(address _gauge) external Governance {
         require(isAlive[_gauge], "gauge already dead");
@@ -339,7 +339,7 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
 
     
     /// @notice Reset the votes of a given TokenID
-    function reset(uint _tokenId) external nonReentrant {
+    function reset(uint256 _tokenId) external nonReentrant {
         _voteDelay(_tokenId);
         require(IVotingEscrow(_ve).isApprovedOrOwner(msg.sender, _tokenId));
         _reset(_tokenId);
@@ -347,13 +347,13 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
         lastVoted[_tokenId] = _epochTimestamp() + 1;
     }
 
-    function _reset(uint _tokenId) internal {
+    function _reset(uint256 _tokenId) internal {
         address[] storage _poolVotes = poolVote[_tokenId];
-        uint _poolVotesCnt = _poolVotes.length;
+        uint256 _poolVotesCnt = _poolVotes.length;
         uint256 _totalWeight = 0;
         uint256 _time = _epochTimestamp();
 
-        for (uint i = 0; i < _poolVotesCnt; i ++) {
+        for (uint256 i = 0; i < _poolVotesCnt; i ++) {
             address _pool = _poolVotes[i];
             uint256 _votes = votes[_tokenId][_pool];
 
@@ -385,14 +385,14 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
     }
 
     /// @notice Recast the saved votes of a given TokenID
-    function poke(uint _tokenId) external nonReentrant {
+    function poke(uint256 _tokenId) external nonReentrant {
         _voteDelay(_tokenId);
         require(IVotingEscrow(_ve).isApprovedOrOwner(msg.sender, _tokenId));
         address[] memory _poolVotes = poolVote[_tokenId];
-        uint _poolCnt = _poolVotes.length;
+        uint256 _poolCnt = _poolVotes.length;
         uint256[] memory _weights = new uint256[](_poolCnt);
 
-        for (uint i = 0; i < _poolCnt; i ++) {
+        for (uint256 i = 0; i < _poolCnt; i ++) {
             _weights[i] = votes[_tokenId][_poolVotes[i]];
         }
 
@@ -405,7 +405,7 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
     /// @param  _tokenId    veNFT tokenID used to vote
     /// @param  _poolVotes   array of LPs addresses to vote  (eg.: [sAMM usdc-usdt   , sAMM busd-usdt, vAMM wbnb-the ,...])
     /// @param  _weights    array of weights for each LPs   (eg.: [10               , 90            , 45             ,...])  
-    function vote(uint _tokenId, address[] calldata _poolVotes, uint256[] calldata _weights) external nonReentrant {
+    function vote(uint256 _tokenId, address[] calldata _poolVotes, uint256[] calldata _weights) external nonReentrant {
         _voteDelay(_tokenId);
         require(IVotingEscrow(_ve).isApprovedOrOwner(msg.sender, _tokenId));
         require(_poolVotes.length == _weights.length);
@@ -413,20 +413,20 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
         lastVoted[_tokenId] = _epochTimestamp() + 1;
     }
     
-    function _vote(uint _tokenId, address[] memory _poolVotes, uint256[] memory _weights) internal {
+    function _vote(uint256 _tokenId, address[] memory _poolVotes, uint256[] memory _weights) internal {
         _reset(_tokenId);
-        uint _poolCnt = _poolVotes.length;
+        uint256 _poolCnt = _poolVotes.length;
         uint256 _weight = IVotingEscrow(_ve).balanceOfNFT(_tokenId);
         uint256 _totalVoteWeight = 0;
         uint256 _totalWeight = 0;
         uint256 _usedWeight = 0;
         uint256 _time = _epochTimestamp();
 
-        for (uint i = 0; i < _poolCnt; i++) {
+        for (uint256 i = 0; i < _poolCnt; i++) {
             _totalVoteWeight += _weights[i];
         }
 
-        for (uint i = 0; i < _poolCnt; i++) {
+        for (uint256 i = 0; i < _poolCnt; i++) {
             address _pool = _poolVotes[i];
             address _gauge = gauges[_pool];
 
@@ -456,44 +456,44 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
 
     /// @notice claim LP gauge rewards
     function claimRewards(address[] memory _gauges) external {
-        for (uint i = 0; i < _gauges.length; i++) {
+        for (uint256 i = 0; i < _gauges.length; i++) {
             IGauge(_gauges[i]).getReward(msg.sender);
         }
     }
 
     /// @notice claim bribes rewards given a TokenID
-    function claimBribes(address[] memory _bribes, address[][] memory _tokens, uint _tokenId) external {
+    function claimBribes(address[] memory _bribes, address[][] memory _tokens, uint256 _tokenId) external {
         require(IVotingEscrow(_ve).isApprovedOrOwner(msg.sender, _tokenId));
-        for (uint i = 0; i < _bribes.length; i++) {
+        for (uint256 i = 0; i < _bribes.length; i++) {
             IBribe(_bribes[i]).getRewardForOwner(_tokenId, _tokens[i]);
         }
     }
 
     /// @notice claim fees rewards given a TokenID
-    function claimFees(address[] memory _fees, address[][] memory _tokens, uint _tokenId) external {
+    function claimFees(address[] memory _fees, address[][] memory _tokens, uint256 _tokenId) external {
         require(IVotingEscrow(_ve).isApprovedOrOwner(msg.sender, _tokenId));
-        for (uint i = 0; i < _fees.length; i++) {
+        for (uint256 i = 0; i < _fees.length; i++) {
             IBribe(_fees[i]).getRewardForOwner(_tokenId, _tokens[i]);
         }
     }
 
     /// @notice claim bribes rewards given an address
     function claimBribes(address[] memory _bribes, address[][] memory _tokens) external {
-        for (uint i = 0; i < _bribes.length; i++) {
+        for (uint256 i = 0; i < _bribes.length; i++) {
             IBribe(_bribes[i]).getRewardForAddress(msg.sender, _tokens[i]);
         }
     }
 
     /// @notice claim fees rewards given an address
     function claimFees(address[] memory _bribes, address[][] memory _tokens) external {
-        for (uint i = 0; i < _bribes.length; i++) {
+        for (uint256 i = 0; i < _bribes.length; i++) {
             IBribe(_bribes[i]).getRewardForAddress(msg.sender, _tokens[i]);
         }
     }    
 
     /// @notice attach a veNFT tokenID to a gauge. This is used for boost farming 
     /// @dev boost not available in Thena. Keep the function in case we need it for future updates. 
-    function attachTokenToGauge(uint tokenId, address account) external {
+    function attachTokenToGauge(uint256 tokenId, address account) external {
         require(isGauge[msg.sender]);
         require(isAlive[msg.sender]); // killed gauges cannot attach tokens to themselves
         if (tokenId > 0) IVotingEscrow(_ve).attach(tokenId);
@@ -503,14 +503,14 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
     
     /// @notice detach a veNFT tokenID to a gauge. This is used for boost farming 
     /// @dev boost not available in Thena. Keep the function in case we need it for future updates. 
-    function detachTokenFromGauge(uint tokenId, address account) external {
+    function detachTokenFromGauge(uint256 tokenId, address account) external {
         require(isGauge[msg.sender]);
         if (tokenId > 0) IVotingEscrow(_ve).detach(tokenId);
         emit Detach(account, msg.sender, tokenId);
     }
 
     /// @notice check if user can vote
-    function _voteDelay(uint _tokenId) internal view {
+    function _voteDelay(uint256 _tokenId) internal view {
         require(block.timestamp > lastVoted[_tokenId] + VOTE_DELAY, "ERR: VOTE_DELAY");
     }
 
@@ -531,7 +531,7 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
         address[] memory _int = new address[](_pool.length);
         address[] memory _ext = new address[](_pool.length);
 
-        uint i = 0;
+        uint256 i = 0;
         for(i; i < _pool.length; i++){
             (_gauge[i], _int[i], _ext[i]) = _createGauge(_pool[i], _gaugeTypes[i]);
         }
@@ -597,7 +597,7 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
         _gauge = IGaugeFactory(_gaugeFactory).createGaugeV2(base, _ve, _pool, address(this), _internal_bribe, _external_bribe, isPair);
      
         // approve spending for $the
-        IERC20(base).approve(_gauge, type(uint).max);
+        IERC20(base).approve(_gauge, type(uint256).max);
 
         // save data
         internal_bribes[_gauge] = _internal_bribe;
@@ -624,12 +624,12 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
     ----------------------------------------------------------------------------- */
 
     /// @notice view the total length of the pools
-    function length() external view returns (uint) {
+    function length() external view returns (uint256) {
         return pools.length;
     }
 
     /// @notice view the total length of the voted pools given a tokenId
-    function poolVoteLength(uint tokenId) external view returns(uint) { 
+    function poolVoteLength(uint256 tokenId) external view returns(uint256) { 
         return poolVote[tokenId].length;
     }
 
@@ -637,7 +637,7 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
         return _factories;
     }
     
-    function factoryLength() external view returns(uint){
+    function factoryLength() external view returns(uint256){
         return _factories.length;
     }
     
@@ -645,29 +645,29 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
         return _gaugeFactories;
     }
     
-    function gaugeFactoriesLength() external view returns(uint) {
+    function gaugeFactoriesLength() external view returns(uint256) {
         return _gaugeFactories.length;
     }
 
-    function weights(address _pool) public view returns(uint) {
-        uint _time = _epochTimestamp();
+    function weights(address _pool) public view returns(uint256) {
+        uint256 _time = _epochTimestamp();
         return weightsPerEpoch[_time][_pool];
     }
 
-    function weightsAt(address _pool, uint _time) public view returns(uint) {
+    function weightsAt(address _pool, uint256 _time) public view returns(uint256) {
         return weightsPerEpoch[_time][_pool];
     }
 
-    function totalWeight() public view returns(uint) {
-        uint _time = _epochTimestamp();
+    function totalWeight() public view returns(uint256) {
+        uint256 _time = _epochTimestamp();
         return totalWeightsPerEpoch[_time];
     }
 
-    function totalWeightAt(uint _time) public view returns(uint) {
+    function totalWeightAt(uint256 _time) public view returns(uint256) {
         return totalWeightsPerEpoch[_time];
     }
 
-    function _epochTimestamp() internal view returns(uint) {
+    function _epochTimestamp() internal view returns(uint256) {
         return IMinter(minter).active_period();
     }
 
@@ -685,10 +685,10 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
     /// @notice notify reward amount for gauge
     /// @dev    the function is called by the minter each epoch. Anyway anyone can top up some extra rewards.
     /// @param  amount  amount to distribute
-    function notifyRewardAmount(uint amount) external {
+    function notifyRewardAmount(uint256 amount) external {
         require(msg.sender == minter);
         _safeTransferFrom(base, msg.sender, address(this), amount);     // transfer the distro in
-        uint _totalWeight = totalWeightAt(_epochTimestamp() - 1 weeks);   // minter call notify after updates active_period, loads votes - 1 week
+        uint256 _totalWeight = totalWeightAt(_epochTimestamp() - 1 weeks);   // minter call notify after updates active_period, loads votes - 1 week
         uint256 _ratio = 0;
 
         if(_totalWeight > 0) _ratio = amount * 1e18 / _totalWeight;     // 1e18 adjustment is removed during claim
@@ -702,9 +702,9 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
     /// @notice notify reward amount for gauge
     /// @dev    the function is called by the minter each epoch. Anyway anyone can top up some extra rewards.
     /// @param  amount  amount to distribute
-    function _notifyRewardAmount(uint amount) external {
+    function _notifyRewardAmount(uint256 amount) external {
         _safeTransferFrom(base, msg.sender, address(this), amount); // transfer the distro in
-        uint _totalWeight = totalWeight();
+        uint256 _totalWeight = totalWeight();
         require(_totalWeight > 0);
         uint256 _ratio = amount * 1e18 / _totalWeight; // 1e18 adjustment is removed during claim
         if (_ratio > 0) {
@@ -719,7 +719,7 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
     /// @param  _gauges  gauge address where to claim the fees 
     /// @dev    the gauge is the owner of the LPs so it has to claim
     function distributeFees(address[] memory _gauges) external {
-        for (uint i = 0; i < _gauges.length; i++) {
+        for (uint256 i = 0; i < _gauges.length; i++) {
             if (isGauge[_gauges[i]] && isAlive[_gauges[i]]){
                 IGauge(_gauges[i]).claimFees();
             }
@@ -732,8 +732,8 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
         
         IMinter(minter).update_period();
 
-        uint x = 0;
-        uint stop = pools.length;
+        uint256 x = 0;
+        uint256 stop = pools.length;
         for (x; x < stop; x++) {
             _distribute(gauges[pools[x]]);
         }
@@ -743,9 +743,9 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
     /// @param  start   start index point of the pools array
     /// @param  finish  finish index point of the pools array
     /// @dev    this function is used in case we have too many pools and gasLimit is reached
-    function distribute(uint start, uint finish) public nonReentrant {
+    function distribute(uint256 start, uint256 finish) public nonReentrant {
         IMinter(minter).update_period();
-        for (uint x = start; x < finish; x++) {
+        for (uint256 x = start; x < finish; x++) {
             _distribute(gauges[pools[x]]);
         }
     }
@@ -754,7 +754,7 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
     /// @dev    this function is used in case some distribution fails
     function distribute(address[] memory _gauges) external nonReentrant {
         IMinter(minter).update_period();
-        for (uint x = 0; x < _gauges.length; x++) {
+        for (uint256 x = 0; x < _gauges.length; x++) {
             _distribute(_gauges[x]);
         }
     }
@@ -762,12 +762,12 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
     /// @notice distribute the emission
     function _distribute(address _gauge) internal {
 
-        uint lastTimestamp = gaugesDistributionTimestamp[_gauge];
-        uint currentTimestamp = _epochTimestamp();
+        uint256 lastTimestamp = gaugesDistributionTimestamp[_gauge];
+        uint256 currentTimestamp = _epochTimestamp();
         if(lastTimestamp < currentTimestamp){
             _updateForAfterDistribution(_gauge); // should set claimable to 0 if killed
 
-            uint _claimable = claimable[_gauge];
+            uint256 _claimable = claimable[_gauge];
 
             // distribute only if claimable is > 0, currentEpoch != lastepoch and gauge is alive
             if (_claimable > 0 && isAlive[_gauge]) {
@@ -790,14 +790,14 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
     
     /// @notice update info for given gauges
     function updateFor(address[] memory _gauges) external {
-        for (uint i = 0; i < _gauges.length; i++) {
+        for (uint256 i = 0; i < _gauges.length; i++) {
             _updateFor(_gauges[i]);
         }
     }
 
     /// @notice update gauge info from start to end
-    function updateForRange(uint start, uint end) public {
-        for (uint i = start; i < end; i++) {
+    function updateForRange(uint256 start, uint256 end) public {
+        for (uint256 i = start; i < end; i++) {
             _updateFor(gauges[pools[i]]);
         }
     }
@@ -815,12 +815,12 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
         uint256 _supplied = weightsPerEpoch[_time][_pool];
 
         if (_supplied > 0) {
-            uint _supplyIndex = supplyIndex[_gauge];
-            uint _index = index; // get global index0 for accumulated distro
+            uint256 _supplyIndex = supplyIndex[_gauge];
+            uint256 _index = index; // get global index0 for accumulated distro
             supplyIndex[_gauge] = _index; // update _gauge current position to global position
-            uint _delta = _index - _supplyIndex; // see if there is any difference that need to be accrued
+            uint256 _delta = _index - _supplyIndex; // see if there is any difference that need to be accrued
             if (_delta > 0) {
-                uint _share = uint(_supplied) * _delta / 1e18; // add accrued difference for each supplied token
+                uint256 _share = uint256(_supplied) * _delta / 1e18; // add accrued difference for each supplied token
                 if (isAlive[_gauge]) {
                     claimable[_gauge] += _share;
                 }
@@ -838,12 +838,12 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
         uint256 _supplied = weightsPerEpoch[_time][_pool];
 
         if (_supplied > 0) {
-            uint _supplyIndex = supplyIndex[_gauge];
-            uint _index = index; // get global index0 for accumulated distro
+            uint256 _supplyIndex = supplyIndex[_gauge];
+            uint256 _index = index; // get global index0 for accumulated distro
             supplyIndex[_gauge] = _index; // update _gauge current position to global position
-            uint _delta = _index - _supplyIndex; // see if there is any difference that need to be accrued
+            uint256 _delta = _index - _supplyIndex; // see if there is any difference that need to be accrued
             if (_delta > 0) {
-                uint _share = uint(_supplied) * _delta / 1e18; // add accrued difference for each supplied token
+                uint256 _share = uint(_supplied) * _delta / 1e18; // add accrued difference for each supplied token
                 if (isAlive[_gauge]) {
                     claimable[_gauge] += _share;
                 }
@@ -876,7 +876,7 @@ contract VoterV3 is IVoter, OwnableUpgradeable, ReentrancyGuardUpgradeable {
 
     /// @notice Fix wrong timestamp of a tokenId
     /// @dev    this is used only if a user weight is saved into the wrong timestamp in weightsPerEpoch [fix 28/04/2023]
-    function forceResetTo(uint _tokenId) external VoterAdmin {
+    function forceResetTo(uint256 _tokenId) external VoterAdmin {
         lastVoted[_tokenId] = _epochTimestamp() - 1 days;
     }
